% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ri.R
\name{calc_RT}
\alias{calc_RT}
\title{Back-calculate Retention Times}
\usage{
calc_RT(ris, alkanesRT, C_num)
}
\arguments{
\item{ris}{A vector of retention indices used to estimate retention times}

\item{alkanesRT}{A vector of retention times of standard alkanes, in descending order}

\item{C_num}{A vector of the numbers of carbons for each of the alkanes}
}
\value{
A vector of expected retention times
}
\description{
This function back-calculates expected retention times given a Van Den Dool and Kratz \href{https://webbook.nist.gov/chemistry/gc-ri/}{retention index}
}
\examples{
alkanes <- data.frame(RT = c(1.88, 2.23, 5.51, 8.05, 10.99,
                             14.10, 17.20, 20.20, 22.90, 25.60,
                             28.10, 30.50, 32.81, 35.22, 37.30),
                      C_num = 6:20)
calc_RT(1007.942, alkanes$RT, alkanes$C_num)

}
\seealso{
\code{\link{calc_RI}}
}
