% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_scale.R
\name{chem_scale}
\alias{chem_scale}
\title{Scaling Functions for Metabolomics}
\usage{
chem_scale(x, center = TRUE, scale = c("auto", "pareto", "range", "vast",
  "level", "none"))
}
\arguments{
\item{x}{a vector}

\item{center}{logical. Do you want to apply centering?}

\item{scale}{choice of scaling functions.  Defaults to autoscaling (dividing by standard deviation). See details for more.}
}
\value{
scaled vector with attributes showing the scaling and centering parameters
}
\description{
Provides additional scaling functions besides autoscaling.  Reviewed in van den Berg et al. 2006.
}
\examples{
x = c(0, 0.1, 0.2, 10)
y = c(1000, 1232, 2022, 4000)

chem_scale(x, center = TRUE, scale = "auto")
chem_scale(y, center = TRUE, scale = "pareto")

}
