\name{chemhelper.update}
\alias{chemhelper.update}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Update the chemhelper package
}
\description{
Downloads and compiles newest version from github.
}
\usage{
chemhelper.update()
}
%- maybe also 'usage' for other objects documented here.

\value{None}

\author{
Jan Stanstrup, \email{stanstrup@gmail.com}
}
