% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_scraping.R
\name{columns_custom}
\alias{columns_custom}
\title{custom tidier for RI tables with custom temp programs}
\usage{
columns_custom(table)
}
\arguments{
\item{table}{}
}
\value{
table
}
\description{
custom tidier for RI tables with custom temp programs
}
