% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_scraping.R
\name{columns_iso}
\alias{columns_iso}
\title{Custom tider for RI tables with isothermal temperature programs}
\usage{
columns_iso(table)
}
\arguments{
\item{table}{}
}
\value{
a table
}
\description{
Custom tider for RI tables with isothermal temperature programs
}
