% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_scraping.R
\name{columns_ramp}
\alias{columns_ramp}
\title{custom tideir for RI tables with ramp temperature programs}
\usage{
columns_ramp(table)
}
\arguments{
\item{table}{}
}
\value{
a table
}
\description{
custom tideir for RI tables with ramp temperature programs
}
