% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_helpers.R
\name{get_loadings}
\alias{get_loadings}
\title{Extract axis loadings from pls object}
\usage{
get_loadings(pls)
}
\arguments{
\item{pls}{a pls object created by \code{\link{opls}}}
}
\value{
a tibble
}
\description{
Provides a wrapper for \code{\link{getLoadingMN}} from the \link\link{ropls} package that returns a tibble rather than a matrix
}
\examples{
\dontrun{
pls.model <- opls(X, Y)
get_loadings(pls.model)
}
}
