% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_scraping.R
\name{get_RI_tables}
\alias{get_RI_tables}
\title{Scrape Retention Indices from NIST}
\usage{
get_RI_tables(cas, type = c("kovats", "linear", "alkane"),
  polarity = c("polar", "non-polar"), temp_prog = c("isothermal", "ramp",
  "custom"))
}
\arguments{
\item{cas}{cas number}

\item{type}{what kind of RI}

\item{polarity}{polar or non-polar}

\item{temp_prog}{what kind of temperature program}
}
\value{
a table
}
\description{
Scrape Retention Indices from NIST
}
\examples{
get_RI_tables("78-70-6", "alkane", "non-polar", "custom")
}
