% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_helpers.R
\name{get_VIP}
\alias{get_VIP}
\title{Extract VIP scores from pls object}
\usage{
get_VIP(pls)
}
\arguments{
\item{pls}{a pls object created by \code{ropls::opls()}}
}
\value{
a tibble
}
\description{
Provides a wrapper for \code{ropls::getVipVn()} that returns a tibble rather than a named numeric vector.
}
\examples{
\dontrun{
pls.model <- opls(X, Y)
get_VIP(pls.model)
}
}
