% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hull.R
\name{hullfn}
\alias{hullfn}
\title{Calculate convex hulls by group.
taken from https://gist.github.com/mbedward/26296da610fd747a9b9f62672f57e9c1.js
Author: Michael Bedward}
\usage{
hullfn(dat, xcol, ycol, groupcol = NULL, hullcol = "hull")
}
\arguments{
\item{dat}{a data frame}

\item{xcol}{column of x data}

\item{ycol}{column of y data}

\item{groupcol}{grouping columns}

\item{hullcol}{I'm not sure yet??}
}
\value{
Returns a new data frame with rows reduced to those for convex hull vertices and a new hull column giving the order in which to plot (e.g. with geom_polygon).
}
\description{
Calculate convex hulls by group.
taken from https://gist.github.com/mbedward/26296da610fd747a9b9f62672f57e9c1.js
Author: Michael Bedward
}
\examples{
\dontrun{
hulls <- hullfn(data, x, y, groupcol = "grp")
}
}
\author{
Michael Bedward
}
