% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ia.R
\name{parse_IA}
\alias{parse_IA}
\title{Parse IonAnalytics CSV files}
\usage{
parse_IA(file)
}
\arguments{
\item{file}{a path to a csv file exported by IonAnalytics}
}
\value{
a string.
}
\description{
Parse IonAnalytics CSV files
}
\examples{
\dontrun{
parse_IA("report.csv")
}

}
