% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_plots.R
\name{pca_plot}
\alias{pca_plot}
\title{Plot PCA models created by \code{ropls::opls()}}
\usage{
pca_plot(ropls_pca, group_var)
}
\arguments{
\item{ropls_pca}{}

\item{group_var}{}
}
\value{
a ggplot object
}
\description{
Plot PCA models created by \code{ropls::opls()}
}
\examples{
\dontrun{
pca_plot(pca, data$treatment)
}
}
