% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_plots.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA models created by \code{ropls::opls()}}
\usage{
plot_pca(ropls_pca, group_var)
}
\arguments{
\item{ropls_pca}{}

\item{group_var}{}
}
\value{
a ggplot object
}
\description{
Plot PCA models created by \code{ropls::opls()}
}
\examples{
\dontrun{
plot_pca(pca, data$treatment)
}
}
