% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_plots.R
\name{plot_pls}
\alias{plot_pls}
\title{Plot PLS regression models produced by \code{ropls::opls()}}
\usage{
plot_pls(ropls_pls)
}
\arguments{
\item{ropls_pls}{a PLS model with a discrete Y variable produced by \code{ropls::opls()}}
}
\value{
a ggplot object
}
\description{
Plot PLS regression models produced by \code{ropls::opls()}
}
\examples{
\dontrun{
plot_pls(pls)
}
}
