% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_scraping.R
\name{read_html_slow}
\alias{read_html_slow}
\title{Read HTML slowly}
\usage{
read_html_slow(x, ...)
}
\arguments{
\item{x}{a URL}

\item{...}{currently unused}
}
\value{
html
}
\description{
Just adds a 5 second rest after \code{read_html()}.  Useful for web-scraping.
}
\examples{
read_html_slow("https://cran.r-project.org/")
}
