% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multvar.R
\name{sim_multvar}
\alias{sim_multvar}
\title{Construct simulated multivariate data sets}
\usage{
sim_multvar(p_uncorvar, p_corvar, p_discr, cov_corvar = 0.5,
  cov_discr = 0.5, diff_discr = 1, group_variance = c(1, 1), N,
  seed = NA)
}
\arguments{
\item{p_uncorvar}{The number of uncorrelated variables to create.}

\item{p_corvar}{The number of correlated, but not associated with group membership, variables to create.}

\item{p_discr}{The number of discriminating variables to create.  The means of these will differ among groups by \code{diff_discr}.}

\item{cov_corvar}{The covariance for the correlated variables}

\item{cov_discr}{The covariance for discriminating variables \emph{within each group} (see details)}

\item{diff_discr}{The difference in means among groups}

\item{group_variance}{A vector the same length as number of groups (currently only 2) that specifies variances for each group.}

\item{N}{Number of observations or rows to simulate}

\item{seed}{A seed for the random number generator to use.  If \code{NA} is supplied, a random seed will be generated.}
}
\value{
A dataframe with one character column for group membership and columns for each data type.
}
\description{
Constructs multivariate data sets with complex structure.  This function allows for a combination of uncorrelated variables, variables that covary, but are not related to group membership, and variables that contribute to discrimination among groups.
}
\details{
This function builds a multivariate dataset with up to three different types of variable group---variables with no covariance (actually cov = 0.01 to help avoid making non-positive-definite matrices), variables with covariance (adjusted using \code{cov_corvar}) but that do not contribute to group separation, and variables that are constructed with means that differ between groups by a total of \code{diff_discr}.  \code{group_variance} applies only to the last two sets of variables and allows specification of different variances (diagonals in the variance-covariance matrix) between groups.

\strong{NOTE:} While \code{cov_discr} supplies a covariance to \code{mvrnorm}, the actual covarince of these variables also depends on \code{diff_discr}.  It is reccomended that the correlation matrix of the final dataset is inspected to ensure you're getting the correlation structure you want.  For example, you may need to reduce \code{cov_discr} relative \code{cov_corvar} to get these two sets of variables to have similar correlations.
}
\examples{

df <- sim_multvar(p_uncorvar = 6, p_corvar = 6, p_discr = 6, N = 40)
library(dplyr)
df \%>\% select(-group) \%>\% as.matrix() \%>\% cor() \%>\% heatmap(Rowv = NA, Colv = NA)
}
