% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_scraping.R
\name{tidy_RItable}
\alias{tidy_RItable}
\title{Tidier for webscraped RI tables}
\usage{
tidy_RItable(tables)
}
\arguments{
\item{tables}{captured by \code{get_RI_tables}}
}
\value{
a single table
}
\description{
Tidier for webscraped RI tables
}
