% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ri.R
\name{VDDK_RI}
\alias{VDDK_RI}
\title{Calculate a single Van Den Dool and Kratz Retention Index}
\usage{
VDDK_RI(rt, alkanesRT, C_num)
}
\arguments{
\item{rt}{The retention time of the compound}

\item{alkanesRT}{A vector of retention times of alkanes, in descending order}

\item{C_num}{A vector of the numbers of carbons for each of the alkanes}
}
\value{
A retention index
}
\description{
Calculate a single Van Den Dool and Kratz Retention Index
}
