% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ri.R
\name{VDDK_RT}
\alias{VDDK_RT}
\title{Calculate a single retention time given a Van Den Dool and Kratz RI}
\usage{
VDDK_RT(ri, alkanesRT, C_num)
}
\arguments{
\item{ri}{The retention index of the compound}

\item{alkanesRT}{A vector of retention times of alkanes, in descending order}

\item{C_num}{A vector of the numbers of carbons for each of the alkanes}
}
\value{
a retention time
}
\description{
Calculate a single retention time given a Van Den Dool and Kratz RI
}
