% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_helpers.R
\name{format_CAS}
\alias{format_CAS}
\title{Format numbers correctly as CAS numbers}
\usage{
format_CAS(x)
}
\arguments{
\item{x}{an integer or character vector}
}
\value{
a character vector of correctly formatted CAS numbers
}
\description{
Format numbers correctly as CAS numbers
}
\examples{
x <- c(58082, 83341, 12345678, "hexanol")
format_CAS(x)
}
