% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_plots.R
\name{plot_plsda}
\alias{plot_plsda}
\title{Plot PLS-DA models produced by \code{ropls::opls()}}
\usage{
plot_plsda(ropls_plsda, annotate = c("caption", "subtitle"))
}
\arguments{
\item{ropls_plsda}{a PLS model with a discrete Y variable produced by \code{ropls::opls()}}
}
\value{
a ggplot object
}
\description{
Plot PLS-DA models produced by \code{ropls::opls()}
}
\examples{
\dontrun{
plot_plsda(plsda)
}
}
