% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ia.R
\name{read_IA}
\alias{read_IA}
\title{Read IonAnalytics CSV files}
\usage{
read_IA(file)
}
\arguments{
\item{file}{a path to a csv file exported by IonAnalytics}
}
\value{
a dataframe
}
\description{
Reads csv files exported from IonAnalytics methods or integration reports.
}
\examples{
\dontrun{
read_IA("report.csv")
}
}
